--- OpenMarket/OpenMarket/Network/URLSessionProtocol.swift
@@ -1,16 +1,0 @@
-//
-//  URLSessionProtocol.swift
-//  OpenMarket
-//
-//  Created by dudu, safari on 2022/05/09.
-//
-
-import Foundation
-
-typealias DataTaskCompletionHandler = (Data?, URLResponse?, Error?) -> Void
-
-protocol URLSessionProtocol {
-    func dataTask(with url: URL, completionHandler: @escaping DataTaskCompletionHandler) -> URLSessionDataTask
-}
-
-extension URLSession: URLSessionProtocol {}
